requirements/base.txt
astroid==1.5.3
autopep8==1.3.2
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.4
django-filter==1.0.4
django-redis==4.8.0
django-rest-swagger==2.1.2
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
idna==2.5
isort==4.2.15
itypes==1.1.0
Jinja2==2.9.6
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mysqlclient==1.3.10
openapi-codec==1.3.2
pycodestyle==2.3.1
PyJWT==1.5.2
pylint==1.7.2
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
pytz==2017.2
redis==2.10.5
requests==2.18.3
simplejson==3.11.1
six==1.10.0
uritemplate==3.0.0
urllib3==1.22
wrapt==1.10.10
